+++ src/main/java/firok/tiths/effect/EffectVoidInfected.java
@@ -10,10 +10,10 @@ import net.minecraft.potion.EffectType;
 @DevUse(isPlaceholder = true)
 public class EffectVoidInfected extends TithsEffect
 {
-	public static final String UUID_MAX_HEALTH = "12CE48F3-561E-4433-A31C-791850821665";
+	public static final String uuidMaxHealth = "12CE48F3-561E-4433-A31C-791850821665";
 	public EffectVoidInfected()
 	{
 		super(EffectType.HARMFUL, 0x4a0964, false);
-		addAttributesModifier(Attributes.MAX_HEALTH, UUID_MAX_HEALTH, -0.3, AttributeModifier.Operation.MULTIPLY_TOTAL);
+		addAttributesModifier(Attributes.MAX_HEALTH, uuidMaxHealth, -0.3, AttributeModifier.Operation.MULTIPLY_TOTAL);
 	}
 }