+++ changelog.md
@@ -2,7 +2,16 @@
 
 > All Time Stamps: GMT+8
 
-## 0.4.0 _2021-10-06 00:00_
- 
- * 配置开发环境 init dev environment
+## 0.4.17.0 22-01-06 17:00
 
+* 新增方块实体 added tile entities
+  * 基座 pedestal tile
+* 新增方块 added blocks
+  * 创造模式基座 creative pedestal block
+  * Motia 水晶基座 motia pedestal block
+  * 焦黑基座 seared pedestal block
+  * 石质基座 stone pedestal block
+
+## 0.4.0 _2021-10-06 00:00_
+
+* 配置开发环境 init dev environment